Update wallet.rs
Made a small change in the "Show your balance" (/balance) command so that it takes an optional input of a user. Now it can show the balance of any user if the input provided or just simply return the author's balance if no input is provided. I hope this works i didn't test it before committing the pull because this is my first time trying out rust!